+++ fuel-tx/src/tests/valid_cases.rs
@@ -41,6 +41,7 @@ pub fn test_params() -> ConsensusParameters {
         Default::default(),
         Default::default(),
         Default::default(),
+        Default::default(),
     )
 }