+++ fuel-tx/src/transaction/consensus_parameters.rs
@@ -28,12 +28,19 @@ pub struct SettingBlockTransactionSizeLimitNotSupported;
 #[cfg(feature = "std")]
 impl std::error::Error for SettingBlockTransactionSizeLimitNotSupported {}
 
+#[derive(Debug, derive_more::Display)]
+#[display("setting block max transactions is not supported")]
+pub struct SettingBlockMaxTransactionsNotSupported;
+#[cfg(feature = "std")]
+impl std::error::Error for SettingBlockMaxTransactionsNotSupported {}
+
 /// A versioned set of consensus parameters.
 #[derive(Clone, Debug, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize)]
 pub enum ConsensusParameters {
     /// Version 1 of the consensus parameters
     V1(ConsensusParametersV1),
     V2(ConsensusParametersV2),
+    V3(ConsensusParametersV3),
 }
 
 #[cfg(feature = "test-helpers")]
@@ -68,9 +75,10 @@ impl ConsensusParameters {
         base_asset_id: AssetId,
         block_gas_limit: u64,
         block_transaction_size_limit: u64,
+        block_max_transactions: u64,
         privileged_address: Address,
     ) -> Self {
-        Self::V2(ConsensusParametersV2 {
+        Self::V3(ConsensusParametersV3 {
             tx_params,
             predicate_params,
             script_params,
@@ -81,6 +89,7 @@ impl ConsensusParameters {
             base_asset_id,
             block_gas_limit,
             block_transaction_size_limit,
+            block_max_transactions,
             privileged_address,
         })
     }
@@ -90,6 +99,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => &params.tx_params,
             Self::V2(params) => &params.tx_params,
+            Self::V3(params) => &params.tx_params,
         }
     }
 
@@ -98,6 +108,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => &params.predicate_params,
             Self::V2(params) => &params.predicate_params,
+            Self::V3(params) => &params.predicate_params,
         }
     }
 
@@ -106,6 +117,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => &params.script_params,
             Self::V2(params) => &params.script_params,
+            Self::V3(params) => &params.script_params,
         }
     }
 
@@ -114,6 +126,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => &params.contract_params,
             Self::V2(params) => &params.contract_params,
+            Self::V3(params) => &params.contract_params,
         }
     }
 
@@ -122,6 +135,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => &params.fee_params,
             Self::V2(params) => &params.fee_params,
+            Self::V3(params) => &params.fee_params,
         }
     }
 
@@ -130,6 +144,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => params.chain_id,
             Self::V2(params) => params.chain_id,
+            Self::V3(params) => params.chain_id,
         }
     }
 
@@ -138,6 +153,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => &params.gas_costs,
             Self::V2(params) => &params.gas_costs,
+            Self::V3(params) => &params.gas_costs,
         }
     }
 
@@ -146,6 +162,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => &params.base_asset_id,
             Self::V2(params) => &params.base_asset_id,
+            Self::V3(params) => &params.base_asset_id,
         }
     }
 
@@ -154,6 +171,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => params.block_gas_limit,
             Self::V2(params) => params.block_gas_limit,
+            Self::V3(params) => params.block_gas_limit,
         }
     }
 
@@ -166,6 +184,20 @@ impl ConsensusParameters {
                 u64::MAX
             }
             Self::V2(params) => params.block_transaction_size_limit,
+            Self::V3(params) => params.block_transaction_size_limit,
+        }
+    }
+
+    /// Get the block max transactions
+    pub fn block_max_transactions(&self) -> u64 {
+        match self {
+            Self::V1(_) | Self::V2(_) => {
+                // In V1 and V2 there was no limit on the transaction size. For the sake
+                // of backwards compatibility we allow for a largest limit
+                // possible.
+                u64::MAX
+            }
+            Self::V3(params) => params.block_max_transactions,
         }
     }
 
@@ -174,6 +206,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => &params.privileged_address,
             Self::V2(params) => &params.privileged_address,
+            Self::V3(params) => &params.privileged_address,
         }
     }
 }
@@ -184,6 +217,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => params.tx_params = tx_params,
             Self::V2(params) => params.tx_params = tx_params,
+            Self::V3(params) => params.tx_params = tx_params,
         }
     }
 
@@ -192,6 +226,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => params.predicate_params = predicate_params,
             Self::V2(params) => params.predicate_params = predicate_params,
+            Self::V3(params) => params.predicate_params = predicate_params,
         }
     }
 
@@ -200,6 +235,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => params.script_params = script_params,
             Self::V2(params) => params.script_params = script_params,
+            Self::V3(params) => params.script_params = script_params,
         }
     }
 
@@ -208,6 +244,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => params.contract_params = contract_params,
             Self::V2(params) => params.contract_params = contract_params,
+            Self::V3(params) => params.contract_params = contract_params,
         }
     }
 
@@ -216,6 +253,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => params.fee_params = fee_params,
             Self::V2(params) => params.fee_params = fee_params,
+            Self::V3(params) => params.fee_params = fee_params,
         }
     }
 
@@ -224,6 +262,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => params.chain_id = chain_id,
             Self::V2(params) => params.chain_id = chain_id,
+            Self::V3(params) => params.chain_id = chain_id,
         }
     }
 
@@ -232,6 +271,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => params.gas_costs = gas_costs,
             Self::V2(params) => params.gas_costs = gas_costs,
+            Self::V3(params) => params.gas_costs = gas_costs,
         }
     }
 
@@ -240,6 +280,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => params.base_asset_id = base_asset_id,
             Self::V2(params) => params.base_asset_id = base_asset_id,
+            Self::V3(params) => params.base_asset_id = base_asset_id,
         }
     }
 
@@ -248,6 +289,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => params.block_gas_limit = block_gas_limit,
             Self::V2(params) => params.block_gas_limit = block_gas_limit,
+            Self::V3(params) => params.block_gas_limit = block_gas_limit,
         }
     }
 
@@ -262,6 +304,24 @@ impl ConsensusParameters {
                 params.block_transaction_size_limit = block_transaction_size_limit;
                 Ok(())
             }
+            Self::V3(params) => {
+                params.block_transaction_size_limit = block_transaction_size_limit;
+                Ok(())
+            }
+        }
+    }
+
+    /// Set the block max number of transactions.
+    pub fn set_block_max_transactions(
+        &mut self,
+        block_max_transactions: u64,
+    ) -> Result<(), SettingBlockMaxTransactionsNotSupported> {
+        match self {
+            Self::V1(_) | Self::V2(_) => Err(SettingBlockMaxTransactionsNotSupported),
+            Self::V3(params) => {
+                params.block_max_transactions = block_max_transactions;
+                Ok(())
+            }
         }
     }
 
@@ -270,6 +330,7 @@ impl ConsensusParameters {
         match self {
             Self::V1(params) => params.privileged_address = privileged_address,
             Self::V2(params) => params.privileged_address = privileged_address,
+            Self::V3(params) => params.privileged_address = privileged_address,
         }
     }
 }
@@ -385,6 +446,69 @@ impl Default for ConsensusParametersV2 {
 impl From<ConsensusParametersV2> for ConsensusParameters {
     fn from(params: ConsensusParametersV2) -> Self {
         Self::V2(params)
+    }
+}
+
+/// A collection of parameters for convenience
+/// The difference with [`ConsensusParametersV2`]:
+/// - `block_max_transactions` has been added.
+#[derive(Clone, Debug, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize)]
+pub struct ConsensusParametersV3 {
+    pub tx_params: TxParameters,
+    pub predicate_params: PredicateParameters,
+    pub script_params: ScriptParameters,
+    pub contract_params: ContractParameters,
+    pub fee_params: FeeParameters,
+    pub chain_id: ChainId,
+    pub gas_costs: GasCosts,
+    pub base_asset_id: AssetId,
+    pub block_gas_limit: u64,
+    pub block_transaction_size_limit: u64,
+    pub block_max_transactions: u64,
+    /// The privileged address(user or predicate) that can perform permissioned
+    /// operations(like upgrading the network).
+    pub privileged_address: Address,
+}
+
+#[cfg(feature = "test-helpers")]
+impl ConsensusParametersV3 {
+    const DEFAULT_BLOCK_MAX_TRANSACTIONS: u64 = u64::MAX;
+    const DEFAULT_BLOCK_TRANSACTION_SIZE_LIMIT: u64 = 126 * 1024;
+
+    /// Constructor for the `ConsensusParameters` with Standard values.
+    pub fn standard() -> Self {
+        Self::standard_with_id(ChainId::default())
+    }
+
+    /// Constructor for the `ConsensusParameters` with Standard values around `ChainId`.
+    pub fn standard_with_id(chain_id: ChainId) -> Self {
+        Self {
+            tx_params: TxParameters::DEFAULT,
+            predicate_params: PredicateParameters::DEFAULT,
+            script_params: ScriptParameters::DEFAULT,
+            contract_params: ContractParameters::DEFAULT,
+            fee_params: FeeParameters::DEFAULT,
+            chain_id,
+            gas_costs: GasCosts::default(),
+            base_asset_id: Default::default(),
+            block_gas_limit: TxParameters::DEFAULT.max_gas_per_tx(),
+            block_transaction_size_limit: Self::DEFAULT_BLOCK_TRANSACTION_SIZE_LIMIT,
+            block_max_transactions: Self::DEFAULT_BLOCK_MAX_TRANSACTIONS,
+            privileged_address: Default::default(),
+        }
+    }
+}
+
+#[cfg(feature = "test-helpers")]
+impl Default for ConsensusParametersV3 {
+    fn default() -> Self {
+        Self::standard()
+    }
+}
+
+impl From<ConsensusParametersV3> for ConsensusParameters {
+    fn from(params: ConsensusParametersV3) -> Self {
+        Self::V3(params)
     }
 }
 
@@ -1036,6 +1160,8 @@ pub mod typescript {
 mod tests {
     use crate::consensus_parameters::{
         ConsensusParametersV2,
+        ConsensusParametersV3,
+        SettingBlockMaxTransactionsNotSupported,
         SettingBlockTransactionSizeLimitNotSupported,
     };
 
@@ -1063,6 +1189,39 @@ mod tests {
             ConsensusParametersV2::default().into();
 
         let result = consensus_params.set_block_transaction_size_limit(0);
+
+        assert!(matches!(result, Ok(())))
+    }
+
+    #[test]
+    fn error_when_setting_block_max_transactions_in_consensus_parameters_v1_v2() {
+        let mut consensus_params: ConsensusParameters =
+            ConsensusParametersV1::default().into();
+
+        let result = consensus_params.set_block_max_transactions(0);
+
+        assert!(matches!(
+            result,
+            Err(SettingBlockMaxTransactionsNotSupported)
+        ));
+
+        let mut consensus_params: ConsensusParameters =
+            ConsensusParametersV2::default().into();
+
+        let result = consensus_params.set_block_max_transactions(0);
+
+        assert!(matches!(
+            result,
+            Err(SettingBlockMaxTransactionsNotSupported)
+        ));
+    }
+
+    #[test]
+    fn ok_when_setting_block_max_transactions_in_consensus_parameters_v3() {
+        let mut consensus_params: ConsensusParameters =
+            ConsensusParametersV3::default().into();
+
+        let result = consensus_params.set_block_max_transactions(0);
 
         assert!(matches!(result, Ok(())))
     }