+++ fuel-vm/src/checked_transaction.rs
@@ -966,6 +966,7 @@ mod tests {
             Default::default(),
             Default::default(),
             Default::default(),
+            Default::default(),
         )
     }