+++ fuel-vm/src/tests/outputs.rs
@@ -167,6 +167,7 @@ fn correct_change_is_provided_for_coin_outputs_create() {
         *context.get_base_asset_id(),
         context.get_block_gas_limit(),
         context.get_block_transaction_size_limit(),
+        context.get_block_max_transactions(),
         *context.get_privileged_address(),
     );
     let create = create