+++ fuel-vm/src/util.rs
@@ -401,6 +401,10 @@ pub mod test_helpers {
             self.consensus_params.block_transaction_size_limit()
         }
 
+        pub fn get_block_max_transactions(&self) -> u64 {
+            self.consensus_params.block_max_transactions()
+        }
+
         pub fn get_privileged_address(&self) -> &Address {
             self.consensus_params.privileged_address()
         }